requirements.txt
# Core dependencies
numpy==1.22.3
matplotlib==3.5.1
networkx==2.8.0
spacy==3.4.1
wikipedia==1.4.0
scikit-learn==1.0.2
pillow==9.1.0

# Visualization
pyvis==0.2.1
seaborn==0.11.2

# NLP dependencies
nltk==3.7

# GUI
tkinter

# Install spaCy model separately: python -m spacy download en_core_web_sm
